Move Coursier related tasks into sbt.coursierint
Ref #4713
